Separate files for script classes
