+++ _bitcoin_common.vcxproj
@@ -223,11 +223,13 @@
     <ClCompile Include="src\rpc\rawtransaction.cpp" />
     <ClCompile Include="src\rpc\server.cpp" />
     <ClCompile Include="src\scheduler.cpp" />
-    <ClCompile Include="src\ScriptID.cpp" />
     <ClCompile Include="src\script\bitcoinconsensus.cpp" />
     <ClCompile Include="src\script\interpreter.cpp" />
     <ClCompile Include="src\script\ismine.cpp" />
     <ClCompile Include="src\script\script.cpp" />
+    <ClCompile Include="src\script\ScriptID.cpp" />
+    <ClCompile Include="src\script\ScriptNum.cpp" />
+    <ClCompile Include="src\script\script_const.cpp" />
     <ClCompile Include="src\script\script_error.cpp" />
     <ClCompile Include="src\script\sigcache.cpp" />
     <ClCompile Include="src\script\sign.cpp" />
@@ -344,11 +346,13 @@
     <ClInclude Include="src\rpc\register.h" />
     <ClInclude Include="src\rpc\server.h" />
     <ClInclude Include="src\scheduler.h" />
-    <ClInclude Include="src\ScriptID.h" />
     <ClInclude Include="src\script\bitcoinconsensus.h" />
     <ClInclude Include="src\script\interpreter.h" />
     <ClInclude Include="src\script\ismine.h" />
     <ClInclude Include="src\script\script.h" />
+    <ClInclude Include="src\script\ScriptID.h" />
+    <ClInclude Include="src\script\ScriptNum.h" />
+    <ClInclude Include="src\script\script_const.h" />
     <ClInclude Include="src\script\script_error.h" />
     <ClInclude Include="src\script\sigcache.h" />
     <ClInclude Include="src\script\sign.h" />
